requirements.txt
streamlit==1.31.1
langchain==0.1.20
langgraph==0.0.39
faiss-cpu==1.7.4
sentence-transformers==2.2.2
pypdf==4.0.1
groq==0.9.0
python-dotenv==1.0.1